pyproject.toml
[project]
name = "bilibili-mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bilibili-api-python>=17.1.4",
    "fastmcp>=0.4.1",
    "loguru>=0.7.3",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
]
